Move logs tests to discover granules and run that test in serial to avoid race conditions
